fix(tests): make the x86 host test build compile the fork DSP headers
The fork's NEON-heavy DSP headers (disperser, sine shaper, multiband,
stereo filters) are reachable from host test TUs via model headers, and
several carry NEON-typed struct members, so include guards can't fix
parsing. Provide a scalar <arm_neon.h> stand-in in the test mocks dirs
instead (supersedes the __ARM_NEON include guards from the previous
commit), covering exactly the intrinsics the codebase uses; the
ARM/QEMU build keeps the real toolchain header.

Also: exclude crash_breadcrumb.cpp from the 32-bit test glob (drags
song.h -> reverb -> Argon, which test targets don't fetch); host
fallback for the qadd asm in disperser.h; std::sinf-style calls in
harm.h -> portable std::sin float overloads (GCC 12 rejects the former);
drop unusable constexpr from LearnedMIDI::equals* (reads a runtime
global; GCC 12 rejects at definition).

MemoryManagerTests (8 tests) and SmallPointerTests (3 tests) now build
and pass locally with g++-12. UnitTests/spec still need GCC 13+ for
<format>/<print> - covered by CI's g++-14.

+++ src/deluge/dsp/harm.h
@@ -210,8 +210,8 @@ struct HarmParams {
 			float w0 = centerFreq * (6.2831853f / 4294967296.0f);
 
 			float Q = 30.0f - (static_cast<float>(hpf - 1) / 126.0f) * 29.5f;
-			float sinw0 = std::sinf(w0);
-			float cosw0 = std::cosf(w0);
+			float sinw0 = std::sin(w0);
+			float cosw0 = std::cos(w0);
 			float alpha = sinw0 / (2.0f * Q);
 			float inv_a0 = 1.0f / (1.0f + alpha);
 
@@ -295,7 +295,7 @@ struct HarmParams {
 			    kHarmPortaMaxRate - (static_cast<float>(porta) / 127.0f) * (kHarmPortaMaxRate - kHarmPortaMinRate);
 			// Compute final value after N samples: freq = target + (current - target) * (1 - rate)^N
 			float retention = 1.0f - portaRate;
-			float retentionN = std::powf(retention, static_cast<float>(buffer.size()));
+			float retentionN = std::pow(retention, static_cast<float>(buffer.size()));
 			currentFreq = targetFreq + (currentFreq - targetFreq) * retentionN;
 		}
 
@@ -305,7 +305,7 @@ struct HarmParams {
 		// Only recompute when frequency changed (sqrtf is ~30 cycles)
 		if (pitchChanged || currentFreq != targetFreq) {
 			float outputHz = currentFreq * (kHarmSampleRate / 4294967296.0f);
-			float pinkScale = (outputHz > 40.0f) ? std::sqrtf(40.0f / outputHz) : 1.0f;
+			float pinkScale = (outputHz > 40.0f) ? std::sqrt(40.0f / outputHz) : 1.0f;
 			cachedPinkQ31 = static_cast<q31_t>(pinkScale * static_cast<float>(ONE_Q31));
 		}
